app/search_service/requirements.txt
# All services
flask==2.3.3
requests==2.31.0
faiss-cpu==1.7.4
numpy==1.25.2